Refactor recipe-ingredient relationship to use join entity
Replaces the many-to-many relationship between Recipes and Ingredients with an explicit RecipeIngredient join entity, allowing for additional properties such as amount, unit, and note. Updates DTOs, services, and database context to reflect this new structure, and adjusts related logic and mappings throughout the backend. Also fixes a type in SpoonacularDtos and improves error logging in SpoonacularService.

+++ backend/DTOs/SpoonacularDtos.cs
@@ -334,7 +334,7 @@ public class RecipeDetailsResponseDto
     public int AggregateLikes { get; set; }
 
     [JsonPropertyName("healthScore")]
-    public int HealthScore { get; set; }
+    public double HealthScore { get; set; }
 
     [JsonPropertyName("creditsText")]
     public string? CreditsText { get; set; }
